update: rename `Canvas/` -> `Main` and `ConnectAppCanvas` -> `ConnectAppPanel`.
