+++ Assets/ConnectApp/Components/CustomSnackBar.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using ConnectApp.canvas;
+using ConnectApp.Components;
 using ConnectApp.constants;
 using Unity.UIWidgets.animation;
 using Unity.UIWidgets.async;
