+++ Assets/ConnectApp/Screens/EventOfflineDetailScreen.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
-using ConnectApp.canvas;
 using ConnectApp.components;
+using ConnectApp.Components;
 using ConnectApp.constants;
 using ConnectApp.models;
 using ConnectApp.Models.ActionModel;
