+++ Assets/ConnectApp/Screens/HistoryArticleScreen.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
-using ConnectApp.canvas;
 using ConnectApp.components;
+using ConnectApp.Components;
 using ConnectApp.constants;
 using ConnectApp.models;
 using ConnectApp.Models.ActionModel;
